# packages/server/src/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Boardie",
    "description": "API to serve the Boardie application"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/me/matches": {
      "post": {
        "summary": "Creates a match for the logged user",
        "tags": [
          "matches"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "boardgameId": {
                    "type": "number"
                  },
                  "location": {
                    "type": "string",
                    "nullable": true
                  },
                  "startedAt": {
                    "type": "string"
                  },
                  "endedAt": {
                    "type": "string",
                    "nullable": true
                  },
                  "notes": {
                    "type": "string",
                    "nullable": true
                  },
                  "players": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "friend": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "type": {
                              "type": "string",
                              "enum": [
                                "ANON_FRIEND",
                                "USER"
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "type"
                          ],
                          "additionalProperties": false,
                          "description": "Generic friend identification data"
                        },
                        "score": {
                          "type": "number",
                          "nullable": true
                        },
                        "isWinner": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "friend",
                        "score",
                        "isWinner"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "required": [
                  "boardgameId",
                  "location",
                  "startedAt",
                  "endedAt",
                  "notes",
                  "players"
                ],
                "additionalProperties": false,
                "description": "Data used to create a match along with its initial players"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created match",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "authorId": {
                      "type": "number",
                      "nullable": true
                    },
                    "boardgameId": {
                      "type": "number"
                    },
                    "location": {
                      "type": "string",
                      "nullable": true
                    },
                    "startedAt": {
                      "type": "string"
                    },
                    "endedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "notes": {
                      "type": "string",
                      "nullable": true
                    },
                    "players": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "friend": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number"
                              },
                              "type": {
                                "type": "string",
                                "enum": [
                                  "ANON_FRIEND",
                                  "USER"
                                ]
                              },
                              "fullName": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "type",
                              "fullName"
                            ],
                            "additionalProperties": false,
                            "description": "Generic friend"
                          },
                          "score": {
                            "type": "number",
                            "nullable": true
                          },
                          "isWinner": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "id",
                          "friend",
                          "score",
                          "isWinner"
                        ],
                        "additionalProperties": false,
                        "description": "Data that represents an existent player"
                      }
                    },
                    "boardgame": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "title": {
                          "type": "string"
                        },
                        "imageUrl": {
                          "type": "string",
                          "nullable": true
                        },
                        "bggId": {
                          "type": "number"
                        },
                        "year": {
                          "type": "number",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "title",
                        "imageUrl",
                        "bggId",
                        "year"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "authorId",
                    "boardgameId",
                    "location",
                    "startedAt",
                    "endedAt",
                    "createdAt",
                    "deletedAt",
                    "notes",
                    "players",
                    "boardgame"
                  ],
                  "additionalProperties": false,
                  "description": "Match data with its players"
                }
              }
            }
          },
          "403": {
            "description": "The logged user is not friends with one of the players"
          },
          "404": {
            "description": "No boardgame with the given id was found"
          }
        }
      },
      "get": {
        "summary": "Gets all the matches for the logged user",
        "tags": [
          "matches"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The matches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "authorId": {
                        "type": "number",
                        "nullable": true
                      },
                      "boardgameId": {
                        "type": "number"
                      },
                      "location": {
                        "type": "string",
                        "nullable": true
                      },
                      "startedAt": {
                        "type": "string"
                      },
                      "endedAt": {
                        "type": "string",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "deletedAt": {
                        "type": "string",
                        "nullable": true
                      },
                      "notes": {
                        "type": "string",
                        "nullable": true
                      },
                      "players": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "friend": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "number"
                                },
                                "type": {
                                  "type": "string",
                                  "enum": [
                                    "ANON_FRIEND",
                                    "USER"
                                  ]
                                },
                                "fullName": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "id",
                                "type",
                                "fullName"
                              ],
                              "additionalProperties": false,
                              "description": "Generic friend"
                            },
                            "score": {
                              "type": "number",
                              "nullable": true
                            },
                            "isWinner": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "id",
                            "friend",
                            "score",
                            "isWinner"
                          ],
                          "additionalProperties": false,
                          "description": "Data that represents an existent player"
                        }
                      },
                      "boardgame": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "title": {
                            "type": "string"
                          },
                          "imageUrl": {
                            "type": "string",
                            "nullable": true
                          },
                          "bggId": {
                            "type": "number"
                          },
                          "year": {
                            "type": "number",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "imageUrl",
                          "bggId",
                          "year"
                        ],
                        "additionalProperties": false,
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "authorId",
                      "boardgameId",
                      "location",
                      "startedAt",
                      "endedAt",
                      "createdAt",
                      "deletedAt",
                      "notes",
                      "players",
                      "boardgame"
                    ],
                    "additionalProperties": false,
                    "description": "Match data with its players"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}